Backing out some size_t and MPI_IN_PLACE changes that were causing problems

+++ src/brains/ForceField.hpp
@@ -132,7 +132,7 @@ namespace OpenMD {
       wildCardAtomTypeName_ = wildCard;
     }
     
-    size_t getNAtomType() {
+    unsigned int getNAtomType() {
       return atomTypeCont_.size();
     }
