+++ src/types/CutoffGroupStamp.hpp
@@ -48,7 +48,7 @@ namespace OpenMD {
   public:
     CutoffGroupStamp();
     int getMemberAt(int index) {return members_[index];}
-    size_t getNMembers() {return members_.size();}
+    int getNMembers() {return members_.size();}
     std::vector<int> getMembers() {return members_;}
     bool setMembers(const std::vector<int>& members) {
       members_ = members;
