+++ src/types/InversionStamp.hpp
@@ -54,7 +54,7 @@ namespace OpenMD {
     
     int getCenter() {return center_;}
     int getSatelliteAt( int index ) {return satellites_.at(index);}
-    size_t getNSatellites() {return satellites_.size();}
+    int getNSatellites() {return satellites_.size();}
     std::vector<int> getSatellites() {return satellites_;}
     void setCenter(int center) { center_ = center; }
     void addSatellite(int sat) {
